Transfer code to respective files
