+++ src/deck.js
@@ -1,0 +1,8 @@
+const createDeck = deck => ({cards: deck});
+
+const countCards = deck => deck.cards.length;
+
+module.exports = {
+  createDeck,
+  countCards
+};
