+++ src/turn.js
@@ -1,0 +1,5 @@
+const evaluateGuess = guess => guess === card.correctAnswer ? 'correct!' : 'incorrect!';
+
+module.exports = {
+  evaluateGuess
+};
