+++ test/Turn-test.js
@@ -1,0 +1,20 @@
+const chai = require('chai');
+const expect = chai.expect;
+
+const { createCard } = require('../src/card');
+const { evaluateGuess } = require('../src/turn');
+
+describe('turn', function() {
+  it('should be a function', function() {
+    expect(evaluateGuess).to.be.a('function');
+  });
+
+  it('should evaluate if a guess is correct or incorrect', function() {
+    const card = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
+    const correctGuess = evaluateGuess('object', card);
+    const incorrectGuess = evaluateGuess('array', card);
+
+    expect(correctGuess).to.deep.equal('correct!');
+    expect(incorrectGuess).to.deep.equal('incorrect!');
+  });
+});
